Add link-count and standalone-node check
